+++ tests/test_config.py
@@ -8,6 +8,7 @@ def test_config_ini_fail():
     with pytest.raises(MinosConfigException):
         instance = MinosConfig(path='./test_fail_config.yaml')
 
+
 def test_config_rest():
     provider_instance = MinosConfig(path='./tests/test_config.yaml')
     assert provider_instance.rest.broker.host == "localhost"
@@ -19,3 +20,14 @@ def test_config_events():
     provider_instance = MinosConfig(path='./tests/test_config.yaml')
     assert provider_instance.events.broker.host == "localhost"
     assert provider_instance.events.broker.port == 9092
+
+
+def test_config_events_database():
+    provider_instance = MinosConfig(path='./tests/test_config.yaml')
+    assert provider_instance.events.database.path == "./tests/local_db.lmdb"
+    assert provider_instance.events.database.name == "database_events_test"
+
+def test_config_commands_database():
+    provider_instance = MinosConfig(path='./tests/test_config.yaml')
+    assert provider_instance.commands.database.path == "./tests/local_db.lmdb"
+    assert provider_instance.commands.database.name == "database_commands_test"